server/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.6.0
ansible==8.2.0
ansible-core==2.15.2
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
Authlib==1.6.2
bitarray==2.9.2
blinker==1.7.0
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==42.0.5
cycler==0.12.1
datasets==3.1.0
decorator==4.4.2
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.115.4
ffmpeg-python==0.2.0
filelock==3.15.3
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==5.0.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
flask-validate-json==0.0.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.11.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
install==1.3.5
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kagglehub==0.3.12
keras==3.6.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.128
libclang==18.1.1
llvmlite==0.43.0
load-dotenv==0.1.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.13.5
networkx==3.3
numba==0.60.0
numpy==1.26.4
oauthlib==2.1.0
openai==1.47.1
openai-whisper==20231117
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.7
packaging==24.0
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.4
pdfreader==0.1.15
pillow==10.3.0
pluggy==1.4.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pypng==0.20220715.0
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.1.0
resolvelib==1.0.1
rich==13.9.4
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.28
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.1
tbb==2021.13.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.2
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.32.0
watchdog==6.0.0
Werkzeug==3.0.1
whisper==1.1.10
wrapt==1.16.0
xxhash==3.5.0
yarl==1.12.1
